+++ CHANGELOG.md
@@ -59,6 +59,7 @@ v1.8.0 [unreleased]
 - [#339](https://github.com/latera/homs/pull/339) Add polling with pagination for claiming table.
 - [#344](https://github.com/latera/homs/pull/344) Add component for rendering created/deadline date.
 - [#345](https://github.com/latera/homs/pull/345) Add task search.
+- [#341](https://github.com/latera/homs/pull/341) Add short tasks list for popup.
 
 v1.7.8 [2019-10-03]
 -------------------